@@ -1,6 +1,6 @@
 [tool.black]
 line-length = 120
-target_version = ['py310']
+target_version = ['py311']
 include = '\.pyi?$'
 exclude = '''
 /(
