@@ -9,7 +9,7 @@ DEBUG = False
 TEMPLATE_DEBUG = False
 
 COMPRESS_ENABLED = True
-COMPRESS_ROOT = "/tmp/cache"  # nosec
+COMPRESS_ROOT = "/var/lib/manti/cache"  # nosec
 
 ALLOWED_HOSTS = ["127.0.0.1"]
 
@@ -18,10 +18,10 @@ BASE_URL = "http://127.0.0.1"
 LOCALE_URLS = {"be": "127.0.0.1", "ru": "127.0.0.1", "en": "127.0.0.1"}
 
 STATIC_URL = "/static/"
-STATIC_ROOT = "/tmp/static"  # nosec
+STATIC_ROOT = "/var/lib/manti/static"  # nosec
 
 MEDIA_URL = "/content/"
-MEDIA_ROOT = "/tmp/content"  # nosec
+MEDIA_ROOT = "/var/lib/manti/content"  # nosec
 
 CACHES = {
     "default": {
